Refactor: Move ItemService and InteractService to Common
