+++ Common/src/main/java/dk/sdu/mmmi/common/services/IInteractService.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.commoninteract;
+package dk.sdu.mmmi.common.services;
 
 import dk.sdu.mmmi.common.data.Entity;
 import dk.sdu.mmmi.common.data.World;
@@ -7,6 +7,6 @@ import dk.sdu.mmmi.common.data.World;
  *
  * @author Jacob
  */
-public interface InteractSPI {
+public interface IInteractService {
     void interact(Entity user, World world);
 }
