+++ Common/src/main/java/dk/sdu/mmmi/common/services/IItemService.java
@@ -1,16 +1,14 @@
-package dk.sdu.mmmi.commonitem;
+package dk.sdu.mmmi.common.services;
 
 import dk.sdu.mmmi.common.data.Entity;
 import dk.sdu.mmmi.common.data.GameData;
 import java.awt.Image;
 
-
 /**
  *
  * @author Jacob
  */
-
-public interface ItemSPI {
+public interface IItemService {
     void useItem(Entity shooter, GameData gameData);
     String getDescription();
     Image getSprite();
