--- CommonItem/src/main/assembly/felix.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<assembly>
- <id>all</id>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <dependencySets>
-    <dependencySet>
-        <useProjectArtifact>false</useProjectArtifact>
-        <outputDirectory>modules</outputDirectory>
-    </dependencySet>
-  </dependencySets>
-  <files>
-    <file>
-      <source>${project.build.directory}/${project.build.finalName}.jar</source>
-      <outputDirectory>modules</outputDirectory>
-    </file>
-    <file>
-      <source>${project.build.directory}/felix.jar</source>
-      <outputDirectory>bin</outputDirectory>
-    </file>
-    <file>
-      <source>${project.build.directory}/config.properties</source>
-      <outputDirectory>conf</outputDirectory>
-    </file>
-  </files>
-</assembly>
