+++ Player/pom.xml
@@ -21,16 +21,6 @@
             <artifactId>Common</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>dk.sdu.mmmi</groupId>
-            <artifactId>CommonInteract</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>dk.sdu.mmmi</groupId>
-            <artifactId>CommonItem</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
     </dependencies>
 
     <build>
