+++ Player/src/main/resources/META-INF/entityprocessor.xml
@@ -4,10 +4,13 @@
     <service>
         <provide interface="dk.sdu.mmmi.common.services.IEntityProcessingService"/>
     </service>
-    <reference bind="setItemService" cardinality="0..1"
-               interface="dk.sdu.mmmi.commonitem.ItemSPI"
-               name="ItemService" policy="dynamic" unbind="removeItemService"/>
+    
     <reference bind="setInteractService" cardinality="0..1"
-               interface="dk.sdu.mmmi.commoninteract.InteractSPI"
-               name="InteractService" policy="dynamic" unbind="removeInteractService"/>
+               interface="dk.sdu.mmmi.common.services.IInteractService"
+               name="IInteractService" policy="dynamic" unbind="removeInteractService"/>
+    
+    <reference bind="setItemService" cardinality="0..1"
+               interface="dk.sdu.mmmi.common.services.IItemService"
+               name="IItemService" policy="dynamic" unbind="removeItemService"/>
+    
 </scr:component>
