+++ Weapon/pom.xml
@@ -33,15 +33,5 @@
             <artifactId>Common</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>CommonItem</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>CommonInteract</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 </project>
