+++ Weapon/src/main/resources/META-INF/entityprocessor.xml
@@ -3,5 +3,7 @@
     <implementation class="dk.sdu.mmmi.weapon.WeaponProcessor"/>
     <service>
         <provide interface="dk.sdu.mmmi.common.services.IEntityProcessingService"/>
+        <provide interface="dk.sdu.mmmi.common.services.IInteractService"/>
+        <provide interface="dk.sdu.mmmi.common.services.IItemService"/>
     </service>
 </scr:component>
