+++ pom.xml
@@ -154,8 +154,6 @@
     <module>Enemy</module>
         
     <module>Player</module>
-    <module>CommonItem</module>
-    <module>CommonInteract</module>
   </modules>
